RDCC-1706: creating new parent response entity with tests
