+++ src/main/java/uk/gov/hmcts/reform/professionalapi/controller/response/ContactInformationResponseWithDxAddress.java
@@ -1,30 +1,21 @@
 package uk.gov.hmcts.reform.professionalapi.controller.response;
 
+import static java.util.stream.Collectors.toList;
+
 import com.fasterxml.jackson.annotation.JsonProperty;
+
+import java.util.List;
+
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 import uk.gov.hmcts.reform.professionalapi.domain.ContactInformation;
 
-@NoArgsConstructor
 @Getter
-public class ContactInformationResponse {
+public class ContactInformationResponseWithDxAddress extends ContactInformationResponse {
 
     @JsonProperty
-    protected String addressLine1;
-    @JsonProperty
-    protected String addressLine2;
-    @JsonProperty
-    protected String addressLine3;
-    @JsonProperty
-    protected String townCity;
-    @JsonProperty
-    protected String county;
-    @JsonProperty
-    protected String country;
-    @JsonProperty
-    protected String postCode;
+    private List<DxAddressResponse> dxAddress;
 
-    public ContactInformationResponse(ContactInformation contactInfo) {
+    public ContactInformationResponseWithDxAddress(ContactInformation contactInfo) {
         this.addressLine1 = contactInfo.getAddressLine1();
         this.addressLine2 = contactInfo.getAddressLine2();
         this.addressLine3 = contactInfo.getAddressLine3();
@@ -32,5 +23,10 @@ public class ContactInformationResponse {
         this.county = contactInfo.getCounty();
         this.country = contactInfo.getCountry();
         this.postCode = contactInfo.getPostCode();
+        this.dxAddress = contactInfo.getDxAddresses()
+                .stream()
+                .map(DxAddressResponse::new)
+                .collect(toList());
     }
+
 }
